change again

+++ Exercise5_2_Korrawit_Y.py
@@ -1,3 +1,3 @@
-Distance = int(input("ระยะทางที่เคลื่อนที่ (km) : "))
-Time = int(input("ระยะเวลา (t) : "))
+Distance = float(input("ระยะทางที่เคลื่อนที่ (km) : "))
+Time = float(input("ระยะเวลา (t) : "))
 print(float(Distance/Time),"Km/h")
